server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Belphin <findolph0003@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.5"
python-decouple = "^3.8"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"